Refactor rules def toward reuse outside RuleCollectionBuilder.vue.
